+++ app/Http/Controllers/TripsController.php
@@ -42,72 +42,90 @@ class TripsController extends Controller
 {
    public function index()
    {
-	    if (Auth::user()->role == '1') {
-	   		//isLoggedIn();
-	   		//Auth();
-			$trips = DB::table('trips')->get();
-			foreach ($trips as $trip) {
-				if($trip->assignments == ""){
-					DB::select( DB::raw("DELETE FROM tripsassignments WHERE tripids = $trip->id") );
-					Trips::find($trip->id)->delete();
+   		if(isset(Auth::user()->role)){                 
+		    if (Auth::user()->role == '1') {
+		   		//isLoggedIn();
+		   		//Auth();
+				$trips = DB::table('trips')->get();
+				foreach ($trips as $trip) {
+					if($trip->assignments == ""){
+						DB::select( DB::raw("DELETE FROM tripsassignments WHERE tripids = $trip->id") );
+						Trips::find($trip->id)->delete();
+					}
 				}
+				return view('trips.index', compact('trips','assignments'));
 			}
-			return view('trips.index', compact('trips','assignments'));
+			else{
+				//return "Hey! dat mag niet!";
+				$error = "1";
+				return view('alert', compact('error'));
+			}
 		}
 		else{
-			//return "Hey! dat mag niet!";
 			$error = "1";
 			return view('alert', compact('error'));
 		}
 	} 	
 	public function wait(){
-		if (Auth::user()->role == '1') {
-			//isLoggedIn();
-			//Auth();
-			$trips = Request::all();
-	  		$newtrip = Trips::create($trips);
-	  		$newtripid = $newtrip->id;
+		if(isset(Auth::user()->role)){  
+			if (Auth::user()->role == '1') {
+				//isLoggedIn();
+				//Auth();
+				$trips = Request::all();
+		  		$newtrip = Trips::create($trips);
+		  		$newtripid = $newtrip->id;
 
-	  		return redirect('/home/tochten/create/'.$newtripid); 
-			//return view('trips.wait', compact('newtripid'));
+		  		return redirect('/home/tochten/create/'.$newtripid); 
+				//return view('trips.wait', compact('newtripid'));
+			}
+			else{
+				//return "Hey! dat mag niet!";
+				$error = "1";
+				return view('alert', compact('error'));
+			}
 		}
 		else{
-			//return "Hey! dat mag niet!";
 			$error = "1";
 			return view('alert', compact('error'));
 		}
 	}
   	public function create($tripid)
   	{
-  	  	if (Auth::user()->role == '1') {
-	      //isLoggedIn();
-	      //Auth();
-	      $trips=Trips::find($tripid);
+  		if(isset(Auth::user()->role)){  
+	  	  	if (Auth::user()->role == '1') {
+		      //isLoggedIn();
+		      //Auth();
+		      $trips=Trips::find($tripid);
 
-		  $assignments = DB::table('assignments')->get();
+			  $assignments = DB::table('assignments')->get();
 
-		  $name = DB::table('tripsassignments')->where('tripids', $tripid)->pluck('assignmentsids');
+			  $name = DB::table('tripsassignments')->where('tripids', $tripid)->pluck('assignmentsids');
 
-		  $actives = DB::table('tripsassignments')->where('tripids', $tripid)->pluck('active');
+			  $actives = DB::table('tripsassignments')->where('tripids', $tripid)->pluck('active');
 
-		  $ids = implode(',', $name);
+			  $ids = implode(',', $name);
 
-		  //return $ids;
+			  //return $ids;
 
-		  if ($ids == ""){
-		  	$assignments = "";
-		  }
-		  else{
-			$assignments = DB::select( DB::raw("SELECT * FROM assignments WHERE id IN($ids)") );
-		  }
-	      foreach($actives as $key => $value) { 
-	  		$assignments[$key]->active = $value; 
-	  	  }
+			  if ($ids == ""){
+			  	$assignments = "";
+			  }
+			  else{
+				$assignments = DB::select( DB::raw("SELECT * FROM assignments WHERE id IN($ids)") );
+			  }
+		      foreach($actives as $key => $value) { 
+		  		$assignments[$key]->active = $value; 
+		  	  }
 
-		  return view('trips.create',compact('assignments','tripid'));
+			  return view('trips.create',compact('assignments','tripid'));
+			}
+			else{
+				//return "Hey! dat mag niet!";
+				$error = "1";
+				return view('alert', compact('error'));
+			}
 		}
 		else{
-			//return "Hey! dat mag niet!";
 			$error = "1";
 			return view('alert', compact('error'));
 		}
@@ -119,39 +137,45 @@ class TripsController extends Controller
   	*/
   	public function store($tripid)
   	{
-  		if (Auth::user()->role == '1') {
-	  		//isLoggedIn();
-	  		//Auth();
-	  		$trip = Request::all($tripid);
-	  		$assignments = DB::table('tripsassignments')->where('tripids', $tripid)->pluck('tripids');
+  		if(isset(Auth::user()->role)){  
+	  		if (Auth::user()->role == '1') {
+		  		//isLoggedIn();
+		  		//Auth();
+		  		$trip = Request::all($tripid);
+		  		$assignments = DB::table('tripsassignments')->where('tripids', $tripid)->pluck('tripids');
 
-	  		$count = count($assignments);
-	  		if ($count < "1"){
-	  			$prevurl = $_SERVER['HTTP_REFERER'];
+		  		$count = count($assignments);
+		  		if ($count < "1"){
+		  			$prevurl = $_SERVER['HTTP_REFERER'];
 
-			    if($prevurl == "http://". $_SERVER['SERVER_NAME'] ."/home/tochten/create/".$tripid){
-			        $prevurl = "create";
-			    }
-			    elseif($prevurl == "http://". $_SERVER['SERVER_NAME'] ."/home/tochten/edit/".$tripid){
-			        $prevurl = "edit";
-			    }
-	  			return view('trips.error',compact('tripid','prevurl'));
-	  		}
-	  		else{
-				$assignments = array_count_values($assignments);
+				    if($prevurl == "http://". $_SERVER['SERVER_NAME'] ."/home/tochten/create/".$tripid){
+				        $prevurl = "create";
+				    }
+				    elseif($prevurl == "http://". $_SERVER['SERVER_NAME'] ."/home/tochten/edit/".$tripid){
+				        $prevurl = "edit";
+				    }
+		  			return view('trips.error',compact('tripid','prevurl'));
+		  		}
+		  		else{
+					$assignments = array_count_values($assignments);
 
-				$assignments = (array_values($assignments));
+					$assignments = (array_values($assignments));
 
-				$assignments =  implode("",$assignments);
-				DB::table('trips')->where('id', $tripid)->update([
-					'tripname' => $trip['tripname'],
-					'assignments' => $assignments,
-			    ]);
-			    return redirect('/home/tochten'); 
+					$assignments =  implode("",$assignments);
+					DB::table('trips')->where('id', $tripid)->update([
+						'tripname' => $trip['tripname'],
+						'assignments' => $assignments,
+				    ]);
+				    return redirect('/home/tochten'); 
+				}
+			}
+	 		else{
+				//return "Hey! dat mag niet!";
+				$error = "1";
+				return view('alert', compact('error'));
 			}
 		}
- 		else{
-			//return "Hey! dat mag niet!";
+		else{
 			$error = "1";
 			return view('alert', compact('error'));
 		}
@@ -164,25 +188,31 @@ class TripsController extends Controller
 	*/
 	public function show($tripid)
 	{	
-		if (Auth::user()->role == '1') {
-		  //isLoggedIn();
-		  //Auth();
+		if(isset(Auth::user()->role)){
+			if (Auth::user()->role == '1') {
+			  //isLoggedIn();
+			  //Auth();
 
-		  $trips = Trips::find($tripid);
-		  $tripname = $trips->tripname;
+			  $trips = Trips::find($tripid);
+			  $tripname = $trips->tripname;
 
-		  $assignments = DB::table('assignments')->get();
+			  $assignments = DB::table('assignments')->get();
 
-		  $name = DB::table('tripsassignments')->where('tripids', $tripid)->pluck('assignmentsids');
+			  $name = DB::table('tripsassignments')->where('tripids', $tripid)->pluck('assignmentsids');
 
-		  $ids = implode(',', $name);
+			  $ids = implode(',', $name);
 
-		  $assignments = DB::select( DB::raw("SELECT * FROM assignments WHERE id IN($ids)") );
+			  $assignments = DB::select( DB::raw("SELECT * FROM assignments WHERE id IN($ids)") );
 
-		  return view('trips.show',compact('assignments','tripid','tripname'));
+			  return view('trips.show',compact('assignments','tripid','tripname'));
+			}
+			else{
+				//return "Hey! dat mag niet!";
+				$error = "1";
+				return view('alert', compact('error'));
+			}
 		}
 		else{
-			//return "Hey! dat mag niet!";
 			$error = "1";
 			return view('alert', compact('error'));
 		}
@@ -197,33 +227,40 @@ class TripsController extends Controller
 	*/
 	public function edit($tripid)
 	{
-		if (Auth::user()->role == '1') {
-		  //isLoggedIn();
-		 // Auth();
-		  $trips=Trips::find($tripid);
-		  $tripname = $trips->tripname;
+		if(isset(Auth::user()->role)){
+			if (Auth::user()->role == '1') {
+			  //isLoggedIn();
+			 // Auth();
+			  $trips=Trips::find($tripid);
+			  $tripname = $trips->tripname;
 
-		  $assignments = DB::table('assignments')->get();
+			  $assignments = DB::table('assignments')->get();
 
-		  $name = DB::table('tripsassignments')->where('tripids', $tripid)->pluck('assignmentsids');
+			  $name = DB::table('tripsassignments')->where('tripids', $tripid)->pluck('assignmentsids');
 
-		  $actives = DB::table('tripsassignments')->where('tripids', $tripid)->pluck('active');
+			  $actives = DB::table('tripsassignments')->where('tripids', $tripid)->pluck('active');
 
-		  $ids = implode(',', $name);
+			  $ids = implode(',', $name);
 
-		  $assignments = DB::select( DB::raw("SELECT * FROM assignments WHERE id IN($ids)") );
+			  $assignments = DB::select( DB::raw("SELECT * FROM assignments WHERE id IN($ids)") );
 
-		  foreach($actives as $key => $value) { 
-		  	$assignments[$key]->active = $value; 
-		  }
+			  foreach($actives as $key => $value) { 
+			  	$assignments[$key]->active = $value; 
+			  }
 
-		  return view('trips.edit',compact('assignments','tripid','tripname'));
+			  return view('trips.edit',compact('assignments','tripid','tripname'));
+			}
+			else{
+				//return "Hey! dat mag niet!";
+				$error = "1";
+				return view('alert', compact('error'));
+			}
 		}
 		else{
-			//return "Hey! dat mag niet!";
 			$error = "1";
 			return view('alert', compact('error'));
 		}
+
 	}
 
 	/**
@@ -240,12 +277,18 @@ class TripsController extends Controller
 	* @return Response
 	*/
 	public function delete($id){
-		if (Auth::user()->role == '1') {
-			$trip = Trips::find($id);
-			return view('trips.delete',compact('trip','id'));
+		if(isset(Auth::user()->role)){
+			if (Auth::user()->role == '1') {
+				$trip = Trips::find($id);
+				return view('trips.delete',compact('trip','id'));
+			}
+			else{
+				//return "Hey! dat mag niet!";
+				$error = "1";
+				return view('alert', compact('error'));
+			}
 		}
 		else{
-			//return "Hey! dat mag niet!";
 			$error = "1";
 			return view('alert', compact('error'));
 		}
@@ -256,15 +299,21 @@ class TripsController extends Controller
 	{
 	  //isLoggedIn();
 	  //Auth();
-	  if (Auth::user()->role == '1'){
-	  	Trips::find($id)->delete();
-	 	$assignments = DB::select( DB::raw("DELETE FROM tripsassignments WHERE tripids = $id") );
-	  	return redirect('/home/tochten');
-	  }
-	  else{
-	  	//return "Hey! dat mag niet!";
-	  	$error = "1";
-		return view('alert', compact('error'));
-	  }
+	  	if(isset(Auth::user()->role)){
+		    if (Auth::user()->role == '1'){
+		  	 	Trips::find($id)->delete();
+		 		$assignments = DB::select( DB::raw("DELETE FROM tripsassignments WHERE tripids = $id") );
+		  		return redirect('/home/tochten');
+		    }
+		    else{
+		  	//return "Hey! dat mag niet!";
+		  		$error = "1";
+				return view('alert', compact('error'));
+		    }
+		}
+		else{
+			$error = "1";
+			return view('alert', compact('error'));
+		}
 	}
 }
